api/requirements.txt
flask
python-dotenv
flask_cors
flask_jwt_extended
flask_restful
flask_sqlalchemy
bcrypt
sendgrid